+++ include/loadimg.h
@@ -1,0 +1,1 @@
+void loadimg(unsigned int kernel_addr);
