+++ main.c
@@ -3,45 +3,12 @@
 #include "include/time.h"
 #include "include/info.h"
 #include "include/lfb.h"
+#include "include/loadimg.h"
 
-int LOADIMG = 0;
 int SPLASH_ON = 0;
+int LOADIMG = 0;
 unsigned int KERNEL_ADDR = 0x100000;
 
-void loadimg(unsigned int kernel_addr)
-{
-    unsigned int size=0;
-
-    // magic number
-    uart_puts("HANK0438\n");
-    uart_puts("Please send the kernel size...\n");
-    // read the kernel's size
-    size=uart_getc();
-    size|=uart_getc()<<8;
-    size|=uart_getc()<<16;
-    size|=uart_getc()<<24;
-
-    // send negative or positive acknowledge
-    if(size<64 || size>1024*1024) {
-        // size error
-        uart_puts("SE");
-        return;
-    }
-    uart_puts("OK");
-
-    uart_puts("Please input the kernel load address (default: 0x80000):\n");
-    uart_puts("Please send the kernel from UART...\n");
-
-
-    uart_puts("Loading kernel at 0x");
-    uart_hex(kernel_addr);
-    uart_puts(" with size 0x");
-    uart_hex(size);
-    uart_puts(" ...\n");
-    
-    // read the kernel
-    while(size--) *(char *)kernel_addr++ = uart_getc();
-}
 
 void main()
 {
@@ -130,18 +97,8 @@ void main()
             continue;
         }
         if (uart_strcmp(user_input, "exc_brk") == 0) {
-            uart_puts("execute `brk #1`\n");
-            asm volatile ("brk #1");
-            uart_puts("done\n");
-            continue;
-        }
-        if (uart_strcmp(user_input, "exc_data") == 0) {
-            uart_puts("execute `bad address access`\n");
-            unsigned int r;
-            // generate a Data Abort with a bad address access
-            r=*((volatile unsigned int*)0xFFFFFFFFFF000000);
-            // make gcc happy about unused variables :-)
-            r++;
+            uart_puts("execute `brk #0`\n");
+            asm volatile ("brk #0");
             uart_puts("done\n");
             continue;
         }
